+++ Runtime/Scripts/EventBus/IEvent.cs
@@ -1,5 +1,12 @@
-﻿namespace RuntimeUtilities.EventBus {
+﻿using System;
+
+namespace RuntimeUtilities.EventBus {
     public interface IEvent {
         
+    }
+    
+    [Serializable]
+    public struct ExampleEvent : IEvent {
+        public int exampleValue;
     }
 }